re-implement useDialog hook and adjust the code accordingly
