+++ lib/dialogs/store.ts
@@ -1,13 +1,12 @@
-import { nanoid } from 'nanoid';
 import { create } from 'zustand';
-import { type Dialog, type DialogWithoutId } from './dialog-schemas';
+import { type Dialog } from './dialog-schemas';
 
 type DialogState = {
   dialogs: Dialog[];
 };
 
 type DialogActions = {
-  openDialog: (dialog: DialogWithoutId) => void;
+  openDialog: (dialog: Dialog) => void;
   closeDialog: (dialogId: string) => void;
 };
 
@@ -15,12 +14,10 @@ type DialogStore = DialogState & DialogActions;
 
 export const useDialogStore = create<DialogStore>((set) => ({
   dialogs: [],
-  openDialog: (dialog) => {
-    const id = nanoid();
-    return set((state) => ({
-      dialogs: [...state.dialogs, { id, ...dialog }],
-    }));
-  },
+  openDialog: (dialog) =>
+    set((state) => ({
+      dialogs: [...state.dialogs, dialog],
+    })),
   closeDialog: (dialogId) =>
     set((state) => ({
       dialogs: state.dialogs.filter((dialog) => dialog.id !== dialogId),
